+++ odbc_tests/tests/CMakeLists.txt
@@ -9,5 +9,4 @@ add_subdirectory(e2e)
 add_subdirectory(integration)
 add_subdirectory(basic_tests)
 add_subdirectory(bindings_tests)
-add_subdirectory(datatype_tests)
 add_subdirectory(odbc-api)
