--- odbc_tests/tests/datatype_tests/CMakeLists.txt
@@ -1,4 +1,0 @@
-cmake_minimum_required(VERSION 4.0)
-
-add_odbc_test(string_tests string_tests.cpp)
-add_odbc_test(number_tests number_tests.cpp real_tests.cpp)
