+++ odbc_tests/tests/e2e/CMakeLists.txt
@@ -37,13 +37,27 @@ add_odbc_test(e2e_types_boolean_conversion_to_c_char types/boolean_conversion_to
 add_odbc_test(e2e_types_boolean_conversion_to_c_integer types/boolean_conversion_to_c_integer.cpp)
 add_odbc_test(e2e_types_boolean_conversion_to_c_real types/boolean_conversion_to_c_real.cpp)
 add_odbc_test(e2e_types_decfloat types/decfloat.cpp)
-add_odbc_test(e2e_types_float types/float.cpp)
-add_odbc_test(e2e_types_float_incompatible_conversions types/float_incompatible_conversions.cpp)
-add_odbc_test(e2e_types_float_integer_valued_conversions types/float_integer_valued_conversions.cpp)
-add_odbc_test(e2e_types_int types/int.cpp)
 add_odbc_test(e2e_types_number types/number.cpp)
-add_odbc_test(e2e_types_number_conversion_to_c_interval types/number_conversion_to_c_interval.cpp)
-add_odbc_test(e2e_types_number_incompatible_conversions types/number_incompatible_conversions.cpp)
+add_odbc_test(e2e_types_int types/int.cpp)
+add_odbc_test(e2e_types_float types/float.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_bit types/fixed_conversion_to_c_bit.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_binary types/fixed_conversion_to_c_binary.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_char types/fixed_conversion_to_c_char.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_integer types/fixed_conversion_to_c_integer.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_numeric types/fixed_conversion_to_c_numeric.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_real types/fixed_conversion_to_c_real.cpp)
+add_odbc_test(e2e_types_fixed_conversion_to_c_interval types/fixed_conversion_to_c_interval.cpp)
+add_odbc_test(e2e_types_fixed_incompatible_conversions types/fixed_incompatible_conversions.cpp)
+add_odbc_test(e2e_types_fixed_treat_decimal_as_int types/fixed_treat_decimal_as_int.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_bit types/real_conversion_to_c_bit.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_char types/real_conversion_to_c_char.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_binary types/real_conversion_to_c_binary.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_default types/real_conversion_to_c_default.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_integer types/real_conversion_to_c_integer.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_numeric types/real_conversion_to_c_numeric.cpp)
+add_odbc_test(e2e_types_real_conversion_to_c_real types/real_conversion_to_c_real.cpp)
+add_odbc_test(e2e_types_real_incompatible_conversions types/real_incompatible_conversions.cpp)
+add_odbc_test(e2e_types_real_integer_valued_conversions types/real_integer_valued_conversions.cpp)
 add_odbc_test(e2e_types_string types/string.cpp)
 add_odbc_test(e2e_types_string_conversion_to_c_char types/string_conversion_to_c_char.cpp)
 add_odbc_test(e2e_types_string_conversion_to_c_integer types/string_conversion_to_c_integer.cpp)
