+++ tests/definitions/odbc/types/fixed_incompatible_conversions.feature
